requirements.txt
linkml>=1.9.6
pyyaml>=6.0
